test: refactoring test codes
