+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "pretest": "standard",
-    "test": "cd test && tap test.js"
+    "test": "cd test && tap ."
   },
   "repository": {
     "type": "git",
